+++ NEWS.md
@@ -5,7 +5,7 @@
 * In `stat_kaplanmeier_pval()`, a `strata` pointing to a character variable will now be coerced to a factor. This enables the calculation of the p value between the different "categories" of a character value, in particular if `ggKAP(color_var)` points to a character variable. (#37)
 * `ggKMMC(sdata = NULL)` is the new default argument, enabling to use the function with observed data only.
 * The legend of `ggKMMC()` do not inform the statistic used in the KMMC. Used to be "Observed mean" whatever the value of `ggKMMC(fun)`.
-
+* New `ggKAP(scale_x_breaks = ggplot2::waiver())`, passed to `ggplot2::scale_x_continuous(breaks)` to enable more granularity in the definition of breaks on the x-axis. Default is set to respect the default behavior of ggplot2. Overrules the existing `ggKAP(scale_x_break_width)` that just defines the width of regular intervals between breaks.
 
 # PMXtte 0.14.2